Remove redundant update() from ofImage.allocate()
Any function calling ofImage.allocate() will surely make some subsequent
call to update the pixels and/or textures. Such a call to update() is
fairly expensive, especially when bUseTexture == true. So it's a shame
to do this twice.

Note, however, that if we do not call update() inside allocate(),
then we should definitely be setting sidth, height, bpp and type here
instead of in update() as these values may well be required in other
places before update() is actually called.

Note: By cutting out the redundant allocations in this and the previous
commit, I cut the time to ofImage.setFromPixels from around 6ms to 3ms,
which made a big difference when trying to play back a jpeg frame
sequence at 60fps.