ofNode::setParent and clearParent have flags to maintain global transform
if you change parent of a node (e.g. clear the parent, or give it a
parent), the node's world transformation would snap to a new position.
This is understandable because all of a sudden the node's local
transformation is relative to a new matrix. Sometimes however, the
desired behaviour is that the node stays where it is in the world,
while it's place in the heirarchy changes. This is very simple to
implement. The ofNode::setParent and clearParent methods now have a
flag (default set to false for backwards compatibilty) to do exactly
this.