simplifies matrix stack
matrix stack & screen setup checked against gl2Renderer in all possible ofOrientations, +-vFlip, found conformant.

Note that using a modelViewOrientation Matrix leads to a lot of extra headache, since you'd have to split model+view matrices to insert the orientation matrix at the right position of the equation.

On GLSL matrix multiplication order:

vec4 vertexP = projection * view * orientation * model * point

This would lead to a lot of extra calculations, and whenever the view changes (i.e. entering an FBO, using a camera), the orientation matrix would have to be re-calculated.

The current way of applying orientation is based on how the GL2 renderer does it, which I consider the reference implementation for visual comparison. This way we can be confident that geometry will be rendered as expected when porting projects to GL3+ =)

Signed-off-by: tgfrerer <tim@poniesandlight.co.uk>