migrate OS X from NSAutoreleasePool to @autoreleasepool
Offending files found via:
$ find . -type f -name '*.h' -o -name '*.cpp' -o -name '*.mm' -o -name '*.m' -exec grep -l "NSAutoreleasePool" {} \;

While it would appear as though a lot of code has changed, really it is largely due to an extra level of indentation required by @autoreleasepool. Additionally, all lines touched should be normalized to use tabs instead of spaces - I'm not a fan of tabs, but I'd rather be consistent than not.

One API change made in ofSystemAlertDialog() to migrate off a mega deprecated NSAlert API.

One file was updated that targets iOS, seemingly missed in #2888.