remove harfbuzz dependency from freetype
As @carthach pointed out, freetype might fail to compile in some cases since the optional dependency to harfbuzz (a utility to parse extended OpenType features) has not been met.

Since we haven't included & compiled harfbuzz in the past (and freetype for homebrew is fine without it) it seems a sensible choice to compile freetype without it.

If we wanted to include Harfbuzz, we'd need to add a dependency for harfbuzz 0.9.19 or newer: <http://sourceforge.net/projects/freetype/files/freetype2/2.5.3/> But it would create a catch22 while compiling, because it,in turn, depends on freetype.

We also have had a flag to exclude harfbuzz from android builds for a while.
I wonder if we should exclude it from iOS builds, as well...

ping @bakercp