update pg to copy directories as directories
+ before this change, the project generator would copy files and
directories using the same method, pretending directories to be files.
on windows, a file as a directory is only copied recursively when this is
specified in the copy parameters. setting this as default could lead to
potentially hazardous consequences, why it's probably better to only copy
directories recursively (since this is the 'natural way' directories are
copied).